Add layout to app context
