+++ layouts/mobile.js
@@ -1,81 +1,21 @@
-import React, {useState, useContext} from 'react'
-import {FileText, Map, BarChart2} from 'react-feather'
+import React, {useContext} from 'react'
 
-import colors from '../styles/colors'
-
-import {AppContext, ThemeContext} from '../pages'
+import {AppContext} from '../pages'
 
 import DateNav from '../components/date-nav'
-import Scrollable from '../components/scrollable'
-import Statistics from '../components/statistics'
-import Informations from '../components/informations'
-import MobileMap from '../components/mobile-map'
 import LayoutSelector from '../components/layout-selector'
 
-const VIEWS = {
-  map: <MobileMap />,
-  stats: (
-    <Scrollable>
-      <Statistics />
-    </Scrollable>
-  ),
-  informations: (
-    <Scrollable>
-      <Informations />
-    </Scrollable>
-  )
-}
-
 const MobilePage = () => {
-  const [selectedView, setSelectedView] = useState('stats')
-
-  const LAYOUTS = {
-    'Vue d’ensemble': VIEWS[selectedView],
-    'Suivi des hospitalisations': (
-      <h1>Suivi des hospitalisations</h1>
-    ),
-    'Suivi des tests': (
-      <h1>
-        Suivi des tests
-      </h1>
-    )
-  }
-  const [selectedLayout, setSelectedLayout] = useState(Object.keys(LAYOUTS)[0])
-
-  const app = useContext(AppContext)
-  const theme = useContext(ThemeContext)
-
-  const handleClick = view => {
-    app.setSelectedLocation(null)
-    setSelectedView(view)
-  }
+  const {selectedLayout} = useContext(AppContext)
 
   return (
     <div className='mobile-page-container'>
-      <div className='mobile-page-header'>
-        <DateNav disabled={selectedView === 'informations'} />
-        <LayoutSelector
-          selected={selectedLayout}
-          layouts={Object.keys(LAYOUTS)}
-          selectLayout={setSelectedLayout}
-        />
+      <div className='mobile-header'>
+        <DateNav />
+        <LayoutSelector />
       </div>
 
-      <Scrollable>
-        {LAYOUTS[selectedLayout]}
-      </Scrollable>
-
-      <div className='view-selector'>
-        <div className={`${selectedView === 'stats' ? 'selected' : ''}`} onClick={() => handleClick('stats')}>
-          <BarChart2 size={32} color={selectedView === 'stats' ? theme.primary : colors.black} />
-        </div>
-        <div className={`${selectedView === 'map' ? 'selected' : ''}`} onClick={() => handleClick('map')}>
-          <Map size={32} color={selectedView === 'map' ? theme.primary : colors.black} />
-        </div>
-        <div className={`${selectedView === 'informations' ? 'selected' : ''}`} onClick={() => handleClick('informations')}>
-          <FileText size={32} color={selectedView === 'informations' ? theme.primary : colors.black} />
-        </div>
-      </div>
+      {selectedLayout.component}
 
       <style jsx>{`
         .mobile-page-container {
@@ -84,28 +24,8 @@ const MobilePage = () => {
           flex: 1;
         }
 
-        .mobile-page-header {
+        .mobile-header {
           z-index: 3;
-        }
-
-        .view-selector {
-          z-index: 1;
-          display: grid;
-          grid-template-columns: 1fr 1fr 1fr;
-          justify-content: center;
-          align-items: center;
-          background-color: #fff;
-          box-shadow: 0 -1px 4px ${colors.lightGrey};
-        }
-
-        .view-selector > div {
-          padding: 0.5em;
-          margin: auto;
-          margin-bottom: -4px;
-        }
-
-        .view-selector > div.selected {
-          border-top: 4px solid ${theme.primary};
         }
       `}</style>
     </div>
